refactor(entity): 将 User 实体类中的 id 类型从 Long 改为 Integer
- 修改了 User 类中的 id 字段类型，以更好地匹配数据库表结构
- 这个改动可能会影响到使用 User 实体类的其他部分，需要相应地进行类型调整

+++ backend/src/main/java/cn/hmck/entity/User.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @TableName("er_user")
 public class User implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Long id;  // 用户ID
+    private Integer id;  // 用户ID
 
     private String username;  // 用户名
